+++ changelog.md
@@ -33,7 +33,7 @@
   - Added new events to the plugin interface:
     - `bytm:dataStoreSerializerLoaded` - Emitted after all memory-cached DataStore instances' data was lazy-loaded.
   - All DataStore instances that have in-memory cached data will now be lazy-loaded after feature initialization is done. Note: lazy-loading starts after `bytm:allReady`, so it's more aimed at lowering initial data access times.
-  - Features can now have tags associated with them, which is another way of filtering them, like when using the new internal function `configSetFeatsWithTags()`.
+  - Features can now have tags associated with them, which is another way of filtering them, like when using the new internal functions `configSetFeatsWithTags()` and `getFeaturesWithTags()`.
   - Updated the dependencies CoreUtils to v3.8.0 and UserUtils to v11.0.0
 
 <div class="pr-link-cont">